pyproject.toml
[tool.poetry]
name = "chainlit"
version = "1.1.306"
keywords = [
    'LLM',
    'Agents',
    'gen ai',
    'chat ui',
    'chatbot ui',
    'openai',
    'copilot',
    'langchain',
    'conversational ai',
]
description = "Build Conversational AI."
authors = ["Chainlit"]
license = "Apache-2.0 license"
repository = "https://github.com/Chainlit/chainlit"
readme = "README.md"
exclude = ["**/frontend/**/*", "**/copilot/**/**/"]
include = ["**/frontend/dist/**/*", "**/copilot/dist/**/*"]

[tool.poetry.scripts]
# command_name = module_for_handler : function_for_handler
chainlit = 'chainlit.cli:cli'

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0.0"
httpx = ">=0.23.0"
literalai = "0.0.607"
dataclasses_json = "^0.5.7"
fastapi = "^0.110.1"
starlette = "^0.37.2"
uvicorn = "^0.25.0"
python-socketio = "^5.11.0"
aiofiles = "^23.1.0"
syncer = "^2.0.3"
asyncer = "^0.0.2"
nest-asyncio = "^1.5.6"
click = "^8.1.3"
tomli = "^2.0.1"
pydantic = ">=1,<3"
python-dotenv = "^1.0.0"
uptrace = "^1.22.0"
watchfiles = "^0.20.0"
filetype = "^1.2.0"
lazify = "^0.4.0"
packaging = "^23.1"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
# Fix https://stackoverflow.com/questions/77582651/why-doesnt-numpy-support-pep517-builds
numpy = [
    { version = "^1.26", python = ">=3.9" },
    { version = "^1.24.4", python = "<3.9" },
]

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
openai = "^1.11.1"
langchain = "^0.1.5"
llama-index = "^0.10.45"
tenacity = "8.3.0"          # pin until fixed https://github.com/langchain-ai/langchain/issues/22972
transformers = "^4.30.1"
matplotlib = "3.7.1"
farm-haystack = "^1.18.0"
plotly = "^5.18.0"
slack_bolt = "^1.18.1"
discord = "^2.3.2"
botbuilder-core = "^4.15.0"

[tool.poetry.group.mypy]
optional = true

[tool.poetry.group.mypy.dependencies]
mypy = "^1.7.1"
types-requests = "^2.31.0.2"
types-aiofiles = "^23.1.0.5"
mypy-boto3-dynamodb = "^1.34.113"

[tool.mypy]
python_version = "3.8"


[[tool.mypy.overrides]]
module = [
    "anthropic",
    "boto3.dynamodb.types",
    "huggingface_hub.inference_api",
    "fastapi_socketio",
    "botbuilder.*",
    "filetype",
    "haystack.*",
    "langflow",
    "lazify",
    "matplotlib.*",                       # remove when 3.8.0 is out, it should export types
    "plotly.*",
    "nest_asyncio",
    "socketio.*",
    "uptrace",
    "syncer",
    "vertexai.language_models",
    "vertexai.preview.generative_models",
    "google.cloud",
]
ignore_missing_imports = true

[tool.poetry.group.custom-data]
optional = true

[tool.poetry.group.custom-data.dependencies]
asyncpg = "^0.29.0"
SQLAlchemy = "^2.0.28"
boto3 = "^1.34.73"
azure-identity = "^1.14.1"
azure-storage-file-datalake = "^12.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
